Stick entries counter in down file
